feat: add install script

+++ script/backup.sh
@@ -12,5 +12,10 @@ CONFIG_BACKUP_DIR="$REPO_DIR/config"
 mkdir -p "$CONFIG_BACKUP_DIR"
 
 # Backup system environment file
-cp /etc/environment "$CONFIG_BACKUP_DIR"
-echo "Config file /etc/environment backed up to $CONFIG_BACKUP_DIR"
+if [[ -f /etc/environment ]]; then
+	cp /etc/environment "$CONFIG_BACKUP_DIR"
+	echo "Config file /etc/environment backed up to $CONFIG_BACKUP_DIR"
+else
+	echo "Error: /etc/environment not found!"
+	exit 1
+fi
